+++ src/V1/ServiceBricks/Api/AdminPolicyApiController.cs
@@ -32,7 +32,6 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpGet]
-        [Route("{storageKey}")]
         [Route("Get/{storageKey}")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
@@ -48,7 +47,6 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpGet]
-        [Route("")]
         [Route("Get")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
@@ -64,6 +62,7 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpGet]
+        [Route("{storageKey}")]
         [Route("GetAsync/{storageKey}")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
@@ -79,6 +78,7 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpGet]
+        [Route("")]
         [Route("GetAsync")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
@@ -94,8 +94,8 @@ namespace ServiceBricks
         /// <param name="dto"></param>
         /// <returns></returns>
         [HttpPut]
-        [Route("")]
         [Route("Update")]
+        [Consumes("application/json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
@@ -110,7 +110,9 @@ namespace ServiceBricks
         /// <param name="dto"></param>
         /// <returns></returns>
         [HttpPut]
+        [Route("")]
         [Route("UpdateAsync")]
+        [Consumes("application/json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
@@ -126,7 +128,8 @@ namespace ServiceBricks
         /// <returns></returns>
         [HttpPut]
         [Route("UpdateAck")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [Consumes("application/json")]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult UpdateAck([FromBody] TDto dto)
@@ -141,7 +144,8 @@ namespace ServiceBricks
         /// <returns></returns>
         [HttpPut]
         [Route("UpdateAckAsync")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [Consumes("application/json")]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> UpdateAckAsync([FromBody] TDto dto, CancellationToken cancellationToken = default)
@@ -155,9 +159,9 @@ namespace ServiceBricks
         /// <param name="dto"></param>
         /// <returns></returns>
         [HttpPost]
-        [Route("")]
         [Route("Create")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [Consumes("application/json")]
+        [ProducesResponseType((int)HttpStatusCode.Created)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult Create([FromBody] TDto dto)
@@ -171,8 +175,10 @@ namespace ServiceBricks
         /// <param name="dto"></param>
         /// <returns></returns>
         [HttpPost]
+        [Route("")]
         [Route("CreateAsync")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [Consumes("application/json")]
+        [ProducesResponseType((int)HttpStatusCode.Created)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> CreateAsync([FromBody] TDto dto, CancellationToken cancellationToken = default)
@@ -187,7 +193,8 @@ namespace ServiceBricks
         /// <returns></returns>
         [HttpPost]
         [Route("CreateAck")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [Consumes("application/json")]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.Created)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult CreateAck([FromBody] TDto dto)
@@ -202,7 +209,8 @@ namespace ServiceBricks
         /// <returns></returns>
         [HttpPost]
         [Route("CreateAckAsync")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [Consumes("application/json")]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.Created)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> CreateAckAsync([FromBody] TDto dto, CancellationToken cancellationToken = default)
@@ -216,9 +224,8 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpDelete]
-        [Route("{storageKey}")]
         [Route("Delete/{storageKey}")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult Delete([FromRoute] string storageKey)
@@ -232,9 +239,8 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpDelete]
-        [Route("")]
         [Route("Delete")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult DeleteFromQuery([FromQuery] string storageKey)
@@ -248,8 +254,9 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpDelete]
+        [Route("{storageKey}")]
         [Route("DeleteAsync/{storageKey}")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> DeleteAsync([FromRoute] string storageKey, CancellationToken cancellationToken = default)
@@ -263,8 +270,9 @@ namespace ServiceBricks
         /// <param name="storageKey"></param>
         /// <returns></returns>
         [HttpDelete]
+        [Route("")]
         [Route("DeleteAsync")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> DeleteFromQueryAsync([FromQuery] string storageKey, CancellationToken cancellationToken = default)
@@ -279,6 +287,7 @@ namespace ServiceBricks
         /// <returns></returns>
         [HttpPost]
         [Route("Query")]
+        [Consumes("application/json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
@@ -294,6 +303,7 @@ namespace ServiceBricks
         /// <returns></returns>
         [HttpPost]
         [Route("QueryAsync")]
+        [Consumes("application/json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
@@ -310,7 +320,6 @@ namespace ServiceBricks
         /// <param name="patchDocument"></param>
         /// <returns></returns>
         [HttpPatch]
-        [Route("{storageKey}")]
         [Route("Patch/{storageKey}")]
         [Consumes("application/json-patch+json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
@@ -330,7 +339,6 @@ namespace ServiceBricks
         /// <param name="patchDocument"></param>
         /// <returns></returns>
         [HttpPatch]
-        [Route("")]
         [Route("Patch")]
         [Consumes("application/json-patch+json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
@@ -351,6 +359,7 @@ namespace ServiceBricks
         /// <param name="patchDocument"></param>
         /// <returns></returns>
         [HttpPatch]
+        [Route("{storageKey}")]
         [Route("PatchAsync/{storageKey}")]
         [Consumes("application/json-patch+json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
@@ -371,6 +380,7 @@ namespace ServiceBricks
         /// <param name="patchDocument"></param>
         /// <returns></returns>
         [HttpPatch]
+        [Route("")]
         [Route("PatchAsync")]
         [Consumes("application/json-patch+json")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
@@ -394,7 +404,7 @@ namespace ServiceBricks
         [HttpPatch]
         [Route("PatchAck/{storageKey}")]
         [Consumes("application/json-patch+json")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult PatchAck(
@@ -413,7 +423,7 @@ namespace ServiceBricks
         [HttpPatch]
         [Route("PatchAck")]
         [Consumes("application/json-patch+json")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override ActionResult PatchAckFromQuery(
@@ -433,7 +443,7 @@ namespace ServiceBricks
         [HttpPatch]
         [Route("PatchAckAsync/{storageKey}")]
         [Consumes("application/json-patch+json")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> PatchAckAsync(
@@ -453,7 +463,7 @@ namespace ServiceBricks
         [HttpPatch]
         [Route("PatchAckAsync")]
         [Consumes("application/json-patch+json")]
-        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(ProblemDetails), (int)HttpStatusCode.BadRequest)]
         [Authorize(Policy = ServiceBricksConstants.SECURITY_POLICY_ADMIN)]
         public override async Task<ActionResult> PatchAckFromQueryAsync(