+++ src/V1/TestDataTypes/src/V1/Tests/TestFiles/Core/TestTestManager.cs
@@ -1130,7 +1130,7 @@ long proputcTicks = 0;
             var appconfig = serviceProvider.GetRequiredService<IConfiguration>();
             var config = new ConfigurationBuilder()
                 .AddConfiguration(appconfig)
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "false" },
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ExposeSystemErrors", "true" }                    
@@ -1148,7 +1148,7 @@ long proputcTicks = 0;
             var appconfig = serviceProvider.GetRequiredService<IConfiguration>();
             var config = new ConfigurationBuilder()
                 .AddConfiguration(appconfig)
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "true" },
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ExposeSystemErrors", "true" }