+++ src/V1/Tests/TestFiles/StorageEfc/ExampleTestManager.cs
@@ -35,7 +35,7 @@ namespace ServiceBricks.Storage.EntityFrameworkCore.Xunit
             var appConfig = serviceProvider.GetRequiredService<IConfiguration>();
             var config = new ConfigurationBuilder()
                 .AddConfiguration(appConfig)
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "false" },
                 })
@@ -52,7 +52,7 @@ namespace ServiceBricks.Storage.EntityFrameworkCore.Xunit
             var appConfig = serviceProvider.GetRequiredService<IConfiguration>();
             var config = new ConfigurationBuilder()
                 .AddConfiguration(appConfig)
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "true" },
                 })
@@ -129,7 +129,7 @@ namespace ServiceBricks.Storage.EntityFrameworkCore.Xunit
             var appConfig = serviceProvider.GetRequiredService<IConfiguration>();
             var config = new ConfigurationBuilder()
                 .AddConfiguration(appConfig)
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "false" },
                 })
@@ -146,7 +146,7 @@ namespace ServiceBricks.Storage.EntityFrameworkCore.Xunit
             var appConfig = serviceProvider.GetRequiredService<IConfiguration>();
             var config = new ConfigurationBuilder()
                 .AddConfiguration(appConfig)
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "true" },
                 })