+++ src/V1/Tests/TestFiles/StorageEfc/StubApiClientTests.cs
@@ -30,7 +30,7 @@ namespace ServiceBricks.Xunit
             var controller = new ExampleApiController(apiservice, apioptions);
             
             var config = new ConfigurationBuilder()
-                .AddInMemoryCollection(new Dictionary<string, string?>
+                .AddInMemoryCollection(new Dictionary<string, string>
                 {
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "false" },
                     { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":DisableAuthentication", "false" },
@@ -56,7 +56,7 @@ namespace ServiceBricks.Xunit
             var controller = new ExampleApiController(apiservice, apioptions);
             
             var config = new ConfigurationBuilder()
-            .AddInMemoryCollection(new Dictionary<string, string?>
+            .AddInMemoryCollection(new Dictionary<string, string>
             {
                             { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "true" },
                             { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":DisableAuthentication", "false" },